Foundation for switching Matchmaking class to using specific listeners: * Matchmaking listeners now inherit from Specific not Global listeners, * Moved all Matchmaking listeners to Matchmaking class, * Removed Matchmaking subclasses, * Removed methods for initializing Matchmaking subclasses, * Removed all reference to Matchmaking subclasses.
