+++ Assets/Scripts/GalaxyManager/Features/Matchmaking.cs
@@ -1,18 +1,13 @@
 ﻿using Galaxy.Api;
 using System.Collections.Generic;
 using UnityEngine;
+using UnityEngine.UI;
 
 public class Matchmaking : MonoBehaviour
 {
     #region Variables
 
     public List<GalaxyID> lobbyList = new List<GalaxyID>();
-
-    public LobbyCreation LobbyCreation;
-    public LobbyBrowsing LobbyBrowsing;
-    public LobbyManagementMainMenu LobbyManagementMainMenu;
-    public LobbyManagementInGame LobbyManagementInGame;
-    public LobbyChat LobbyChat;
 
     // Variables for storing lobbies data
     private GalaxyID currentLobbyID = null;
@@ -26,77 +21,7 @@ public class Matchmaking : MonoBehaviour
 
     #endregion
 
-    #region Behaviours
-
-    void OnDisable()
-    {
-        ShutdownAllMatchmakingClasses();
-    }
-
-    #endregion
-
     #region Methods
-
-    /* Following methods are used to start and shutdown each and every GalaxyManager feature class separately
-    Note: Each class closes its own listeners whyn disabled */
-    public void StartLobbyCreation()
-    {
-        if (LobbyCreation == null) LobbyCreation = gameObject.AddComponent<LobbyCreation>();
-    }
-
-    public void ShutdownLobbyCreation()
-    {
-        if (LobbyCreation != null) Destroy(LobbyCreation);
-    }
-
-    public void StartLobbyBrowsing()
-    {
-        if (LobbyBrowsing == null) LobbyBrowsing = gameObject.AddComponent<LobbyBrowsing>();
-    }
-
-    public void ShutdownLobbyBrowsing()
-    {
-        if (LobbyBrowsing != null) Destroy(LobbyBrowsing);
-    }
-
-    public void StartLobbyManagementMainMenu()
-    {
-        if (LobbyManagementMainMenu == null) LobbyManagementMainMenu = gameObject.AddComponent<LobbyManagementMainMenu>();
-    }
-
-    public void ShutdownLobbyManagementMainMenu()
-    {
-        if (LobbyManagementMainMenu != null) Destroy(LobbyManagementMainMenu);
-    }
-
-    public void StartLobbyManagementInGame()
-    {
-        if (LobbyManagementInGame == null) LobbyManagementInGame = gameObject.AddComponent<LobbyManagementInGame>();
-    }
-
-    public void ShutdownLobbyManagementInGame()
-    {
-        if (LobbyManagementInGame != null) Destroy(LobbyManagementInGame);
-    }
-
-    public void StartLobbyChat()
-    {
-        if (LobbyChat == null) LobbyChat = gameObject.AddComponent<LobbyChat>();
-    }
-
-    public void ShutdownLobbyChat()
-    {
-        if (LobbyChat != null) Destroy(LobbyChat);
-    }
-
-    public void ShutdownAllMatchmakingClasses()
-    {
-        ShutdownLobbyCreation();
-        ShutdownLobbyBrowsing();
-        ShutdownLobbyManagementMainMenu();
-        ShutdownLobbyManagementInGame();
-        ShutdownLobbyChat();
-    }
 
     /* Requests list of available lobbies
     Note: Private lobbies will not be retrieved */
@@ -384,6 +309,436 @@ public class Matchmaking : MonoBehaviour
         string message = null;
         GalaxyInstance.Matchmaking().GetLobbyMessage(lobbyID, messageID, ref senderID, out message);
         return message;
+    }
+
+    #endregion
+
+    #region Lobby browsing listeners
+
+    private LobbyListListenerBrowsing lobbyListListenerBrowsing;
+    private LobbyDataListenerBrowsing lobbyDataListenerBrowsing;
+    private LobbyEnteredListenerBrowsing lobbyEnteredListenerBrowsing;
+
+    /* Informs about the event of retrieving the list of available lobbies
+    Callback to methods:
+    Matchmaking.RequestLobbyList(bool allowFull = false) */
+    private class LobbyListListenerBrowsing : ILobbyListListener
+    {
+        Matchmaking matchmaking = GalaxyManager.Instance.Matchmaking;
+        List<GalaxyID> lobbyList = GalaxyManager.Instance.Matchmaking.lobbyList;
+
+        public override void OnLobbyList(uint count, LobbyListResult result)
+        {
+            if (result == LobbyListResult.LOBBY_LIST_RESULT_SUCCESS)
+            {
+                LobbyListRetrieved(count);
+            }
+            else {
+                Debug.LogWarning("OnLobbyList failure reason: " + result);
+            }
+        }
+
+        private void LobbyListRetrieved(uint count)
+        {
+            Debug.Log(count + " lobbies OnLobbyList");
+            if (count == 0)
+            {
+                GameObject.Find("OnlineBrowserScreen").GetComponent<OnlineBrowserController>().DisplayLobbyList(lobbyList);
+            }
+            else
+            {
+                for (uint i = 0; i < count; i++)
+                {
+                    GalaxyID lobbyID = GalaxyInstance.Matchmaking().GetLobbyByIndex(i);
+                    lobbyList.Add(lobbyID);
+                    Debug.Log("Requesting lobby data for lobby " + i + " with lobbyID " + lobbyID.ToString());
+                    matchmaking.RequestLobbyData(lobbyID);
+                }
+            }
+        }
+
+    }
+
+    /* Informs about the event of receiving specified lobby or lobby member data
+    Callback to methods:
+    Matchmaking.RequestLobbyData(GalaxyID lobbyID)
+    Matchmaking.SetLobbyData(GalaxyID lobbyID, string key, string value)
+    Matchmaking.SetLobbyMemberData(string key, string value) */
+    private class LobbyDataListenerBrowsing : ILobbyDataListener
+    {
+        List<GalaxyID> lobbyList = GalaxyManager.Instance.Matchmaking.lobbyList;
+        public uint lobbiesWithDataRetrievedCount = 0;
+
+        public override void OnLobbyDataUpdated(GalaxyID lobbyID, GalaxyID memberID)
+        {
+            Debug.Log("LobbyID: " + lobbyID + "\nMemberID: " + memberID);
+            if (memberID == new GalaxyID(0)) LobbyDataUpdated();
+        }
+
+        void LobbyDataUpdated()
+        {
+            lobbiesWithDataRetrievedCount++;
+            Debug.Log("Data retrieved for " + lobbiesWithDataRetrievedCount + " lobbies out of " + lobbyList.Count);
+            if (lobbiesWithDataRetrievedCount >= lobbyList.Count)
+            {
+                GameObject.Find("OnlineBrowserScreen").GetComponent<OnlineBrowserController>().DisplayLobbyList(lobbyList);
+                lobbiesWithDataRetrievedCount = 0;
+                lobbyList.Clear();
+                lobbyList.TrimExcess();
+            }
+        }
+
+    }
+
+    /* Informs about the event of entering a lobby
+    Callback for methods:
+    Matchmaking.JoinLobby(GalaxyID lobbyID) */
+    private class LobbyEnteredListenerBrowsing : ILobbyEnteredListener
+    {
+        Matchmaking matchmaking = GalaxyManager.Instance.Matchmaking;
+
+        public override void OnLobbyEntered(GalaxyID lobbyID, LobbyEnterResult _result)
+        {
+            switch (_result)
+            {
+                case LobbyEnterResult.LOBBY_ENTER_RESULT_SUCCESS:
+                    LobbyEntered(lobbyID);
+                break;
+                case LobbyEnterResult.LOBBY_ENTER_RESULT_LOBBY_DOES_NOT_EXIST:
+                    LobbyEnteringError("Lobby does not exist");
+                break;
+                case LobbyEnterResult.LOBBY_ENTER_RESULT_LOBBY_IS_FULL:
+                    LobbyEnteringError("Lobby is full");
+                break;
+                case LobbyEnterResult.LOBBY_ENTER_RESULT_ERROR:
+                    LobbyEnteringError("Unspecified error");
+                break;
+            }
+        }
+
+        public void LobbyEntered(GalaxyID lobbyID)
+        {
+            GalaxyManager.Instance.Matchmaking.CurrentLobbyID = lobbyID;
+            GalaxyManager.Instance.Matchmaking.LobbyOwnerID = matchmaking.GetLobbyOwner(lobbyID);
+            matchmaking.SetLobbyMemberData("state", "notReady");
+            GameObject.Find("MainMenu").GetComponent<MainMenuController>().SwitchMenu(MainMenuController.MenuEnum.OnlineWait);
+        }
+
+        public void LobbyEnteringError(string reason)
+        {
+            GameObject.Find("MainMenu").GetComponent<MainMenuController>().SwitchMenu(MainMenuController.MenuEnum.OnlineBrowser);
+            GameObject.Find("PopUps").GetComponent<PopUps>().MenuCouldNotJoin(reason);
+        }
+
+    }
+
+    #endregion
+
+    #region Lobby chat listeners
+
+    public LobbyMessageListenerChat lobbyMessageListener;
+    public List<Dictionary<string, string>> lobbyMessageHistory = new List<Dictionary<string, string>>();
+
+	// Lobby message listener
+    public class LobbyMessageListenerChat : ILobbyMessageListener
+    {
+        Matchmaking matchmaking = GalaxyManager.Instance.Matchmaking;
+        string message = null;
+
+        public override void OnLobbyMessageReceived(GalaxyID lobbyID, GalaxyID senderID, uint messageID, uint messageLength)
+        {
+            Dictionary<string, string> messageAndSenderDict = new Dictionary<string, string>();
+            try
+            {
+                Debug.Log("Lobby " + lobbyID + " Sender " + senderID + " message " + message);
+                message = matchmaking.GetLobbyMessage(GalaxyManager.Instance.Matchmaking.CurrentLobbyID, ref senderID, messageID);
+                messageAndSenderDict.Add("sender", GalaxyManager.Instance.Friends.GetFriendPersonaName(senderID));
+                messageAndSenderDict.Add("message", message);
+                matchmaking.lobbyMessageHistory.Add(messageAndSenderDict);
+                Debug.Log("New message from " + GalaxyManager.Instance.Friends.GetFriendPersonaName(senderID) + " to lobbyID " + lobbyID + ": " + message);
+                if (GameManager.Instance != null) ((Online2PlayerGameManager)GameManager.Instance).PopChatPrompt();
+            }
+            catch (GalaxyInstance.Error e)
+            {
+                Debug.LogWarning(e);
+            }
+        }
+    }
+
+    #endregion
+
+    #region Lobby creation listeners
+
+    private LobbyCreatedListenerCreation lobbyCreatedListener;
+
+    /* Informs about the event of creating a lobby
+    Callback to methods:
+    Matchmaking.CreateLobby(LobbyType lobbyType, uint maxMembers, bool joinable, LobbyTopologyType lobbyTopologyType) */
+    private class LobbyCreatedListenerCreation : ILobbyCreatedListener
+    {
+        Matchmaking matchmaking = GalaxyManager.Instance.Matchmaking;
+
+        public override void OnLobbyCreated(GalaxyID lobbyID, LobbyCreateResult _result)
+        {
+            switch (_result)
+            {
+                case LobbyCreateResult.LOBBY_CREATE_RESULT_SUCCESS:
+                    LobbyCreated(lobbyID);
+                break;
+                case LobbyCreateResult.LOBBY_CREATE_RESULT_ERROR:
+                    Timeout();
+                break;
+            }
+        }
+
+        private void LobbyCreated(GalaxyID lobbyID)
+        {
+            matchmaking.SetLobbyData(lobbyID, "name", matchmaking.lobbyName);
+            matchmaking.SetLobbyData(lobbyID, "state", "notReady");
+
+            matchmaking.CurrentLobbyID = lobbyID;
+            matchmaking.LobbyOwnerID = GalaxyInstance.Matchmaking().GetLobbyOwner(lobbyID);
+            matchmaking.SetLobbyMemberData("state", "notReady");
+            GameObject.Find("MainMenu").GetComponent<MainMenuController>().SwitchMenu(MainMenuController.MenuEnum.OnlineWait);
+        }
+
+        private void Timeout()
+        {
+            GameObject.Find("MainMenu").GetComponent<MainMenuController>().SwitchMenu(MainMenuController.MenuEnum.OnlineCreate);
+            GameObject.Find("PopUps").GetComponent<PopUps>().MenuCouldNotCreate();
+        }
+
+    }
+
+    #endregion
+
+    #region Lobby managment in game listeners
+
+    private LobbyDataListenerInGame lobbyDataListenerInGame;
+    private LobbyLeftListenerInGame lobbyLeftListenerInGame;
+    private LobbyMemberStateListenerInGame lobbyMemberStateListenerInGame;
+
+    /* Informs about the event of receiving specified lobby or lobby member data
+    Callback to methods:
+    Matchmaking.RequestLobbyData(GalaxyID lobbyID)
+    Matchmaking.SetLobbyData(GalaxyID lobbyID, string key, string value)
+    Matchmaking.SetLobbyMemberData(string key, string value) */
+    private class LobbyDataListenerInGame : ILobbyDataListener
+    {
+        Matchmaking matchmaking = GalaxyManager.Instance.Matchmaking;
+
+        public override void OnLobbyDataUpdated(GalaxyID lobbyID, GalaxyID memberID)
+        {
+            Debug.Log("LobbyID: " + lobbyID + "\nMemberID: " + memberID);
+            if (memberID != new GalaxyID(0)) LobbyMemberDataUpdated(lobbyID, memberID);
+        }
+
+        void LobbyMemberDataUpdated(GalaxyID lobbyID, GalaxyID memberID)
+        {
+            if (AllMembersGo(lobbyID, memberID))
+            {
+                matchmaking.SetLobbyData(lobbyID, "state", "go");
+                GameObject.Find("GameManager").GetComponent<Online2PlayerGameManager>().enabled = true;
+                GameObject.Find("PopUps").GetComponent<PopUps>().ClosePopUps();
+            }
+        }
+
+        bool AllMembersGo(GalaxyID lobbyID, GalaxyID memberID)
+        {
+            uint go = 0;
+
+            // Check how many players are in game
+            for (uint i = 0; i < 2; i++)
+            {
+                if (matchmaking.GetLobbyMemberData(matchmaking.GetLobbyMemberByIndex(lobbyID, i), "state") == "go")
+                {
+                    go++;
+                }
+            }
+
+            return (go == 2) ? true : false;
+
+        }
+
+    }
+
+    /* Informs about the event of leaving a lobby 
+    Callback for methods: 
+    Matchmaking.LeaveLobby() */
+    private class LobbyLeftListenerInGame : ILobbyLeftListener
+    {
+        Matchmaking matchmaking = GalaxyManager.Instance.Matchmaking;
+        public override void OnLobbyLeft(GalaxyID lobbyID, bool ioFailure)
+        {
+            if (!ioFailure)
+            {
+                if (!matchmaking.leftOnMyOwn)
+                {
+                    if (!GameObject.Find("Online2PlayerGameEnd")) HostLeftLobby();
+                }
+                LobbyLeft(lobbyID);
+            }
+            else
+            {
+                Debug.LogWarning("OnLobbyLeft failure " + ioFailure);
+            }
+        }
+
+        void HostLeftLobby()
+        {
+            Debug.Log("Host left the lobby");
+            GameObject.Find("PopUps").GetComponent<PopUps>().ClosePopUps();
+            GameObject.Find("PopUps").GetComponent<PopUps>().GameHostLeftLobby();
+            GameManager.Instance.GameFinished = true;
+        }
+
+        void LobbyLeft(GalaxyID lobbyID)
+        {
+            matchmaking.CurrentLobbyID = null;
+            matchmaking.LobbyOwnerID = null;
+            matchmaking.leftOnMyOwn = false;
+            GalaxyManager.Instance.ShutdownNetworking();
+            Debug.Log("Lobby " + lobbyID + " left");
+        }
+
+    }
+
+    /* Informs about the event of lobby member state change */
+    private class LobbyMemberStateListenerInGame : ILobbyMemberStateListener
+    {
+        public override void OnLobbyMemberStateChanged(GalaxyID lobbyID, GalaxyID memberID, LobbyMemberStateChange memberStateChange)
+        {
+            Debug.Log(string.Format("OnLobbyMemberStateChanged lobbyID: {0} memberID: {1} change: {2}", lobbyID, memberID, memberStateChange));
+            if (memberStateChange != LobbyMemberStateChange.LOBBY_MEMBER_STATE_CHANGED_ENTERED)
+            {
+                if (!GameObject.Find("Online2PlayerGameEnd")) ClientLeftLobby();
+            }
+        }
+
+        private void ClientLeftLobby()
+        {
+            Debug.Log("Client left the lobby.");
+            GameObject.Find("PopUps").GetComponent<PopUps>().ClosePopUps();
+            GameObject.Find("PopUps").GetComponent<PopUps>().GameClientLeftLobby();
+            GameManager.Instance.GameFinished = true;
+        }
+
+    }
+
+    #endregion
+
+    #region Lobby managment main menu listeners
+
+    private LobbyLeftListenerMainMenu lobbyLeftListenerMainMenu;
+    private LobbyDataListenerMainMenu lobbyDataListenerMainMenu;
+
+    /* Informs about the event of receiving specified lobby or lobby member data
+    Callback to methods:
+    Matchmaking.RequestLobbyData(GalaxyID lobbyID)
+    Matchmaking.SetLobbyData(GalaxyID lobbyID, string key, string value)
+    Matchmaking.SetLobbyMemberData(string key, string value) */
+    private class LobbyDataListenerMainMenu : ILobbyDataListener
+    {
+        Matchmaking matchmaking = GalaxyManager.Instance.Matchmaking;
+
+        public override void OnLobbyDataUpdated(GalaxyID lobbyID, GalaxyID memberID)
+        {
+            Debug.Log("LobbyID: " + lobbyID + "\nMemberID: " + memberID);
+            if (memberID == new GalaxyID(0))
+            {
+                LobbyDataUpdatedReady(lobbyID);
+                LobbyDataUpdatedSteady(lobbyID);
+            }
+            else
+            {
+                LobbyMemberDataUpdatedReady(lobbyID);
+            }
+        }
+
+        void LobbyDataUpdatedReady(GalaxyID lobbyID)
+        {
+            if (matchmaking.GetLobbyData(lobbyID, "state") == "ready" && GalaxyManager.Instance.MyGalaxyID == matchmaking.LobbyOwnerID)
+            {
+                GameObject.Find("OnlineWaitScreen").GetComponent<OnlineWaitController>().startGameButton.GetComponent<Button>().interactable = true;
+            }
+            else
+            {
+                GameObject.Find("OnlineWaitScreen").GetComponent<OnlineWaitController>().startGameButton.GetComponent<Button>().interactable = false;
+            }
+        }
+
+        void LobbyDataUpdatedSteady(GalaxyID lobbyID)
+        {
+            if (matchmaking.GetLobbyData(lobbyID, "state") == "steady")
+            {
+                Debug.Assert(matchmaking.GetLobbyMemberData(GalaxyManager.Instance.MyGalaxyID, "state") == "ready");
+                matchmaking.SetLobbyMemberData("state", "steady");
+                SceneController.Instance.LoadScene(SceneController.SceneName.Online2PlayerGame, true);
+            }
+        }
+
+        void LobbyMemberDataUpdatedReady(GalaxyID lobbyID)
+        {
+            if (GalaxyManager.Instance.MyGalaxyID == matchmaking.LobbyOwnerID)
+            {
+                matchmaking.SetLobbyData(lobbyID, "state", AllMembersReady(lobbyID) ? "ready" : "notReady");
+            }
+        }
+
+        bool AllMembersReady(GalaxyID lobbyID)
+        {
+            uint ready = 0;
+            uint lobbyMembersCount = matchmaking.GetNumLobbyMembers(lobbyID);
+            // Checks how many players are ready
+            for (uint i = 0; i < lobbyMembersCount; i++)
+            {
+                if (matchmaking.GetLobbyMemberData(matchmaking.GetLobbyMemberByIndex(lobbyID, i), "state") == "ready")
+                {
+                    ready++;
+                }
+            }
+            return (ready == 2) ? true : false;
+        }
+    }
+
+    /* Informs about the event of leaving a lobby 
+    Callback for methods: 
+    Matchmaking.LeaveLobby() */
+    private class LobbyLeftListenerMainMenu : ILobbyLeftListener
+    {
+        Matchmaking matchmaking = GalaxyManager.Instance.Matchmaking;
+        public override void OnLobbyLeft(GalaxyID lobbyID, bool ioFailure)
+        {
+            if (!ioFailure)
+            {
+                if (!matchmaking.leftOnMyOwn)
+                {
+                    HostLeftLobby();
+                }
+                LobbyLeft(lobbyID);
+            }
+            else
+            {
+                Debug.LogWarning("OnLobbyLeft failure");
+            }
+        }
+
+        void HostLeftLobby()
+        {
+            Debug.Log("Host left the lobby");
+            GameObject.Find("PopUps").GetComponent<PopUps>().ClosePopUps();
+            GameObject.Find("PopUps").GetComponent<PopUps>().MenuHostLeftLobby();
+            GameObject.Find("MainMenu").GetComponent<MainMenuController>().SwitchMenu(MainMenuController.MenuEnum.Online);
+        }
+
+        void LobbyLeft(GalaxyID lobbyID)
+        {
+            matchmaking.CurrentLobbyID = null;
+            matchmaking.LobbyOwnerID = null;
+            matchmaking.leftOnMyOwn = false;
+            Debug.Log("Lobby " + lobbyID + " left");
+        }
+
     }
 
     #endregion
