--- Assets/Scripts/GalaxyManager/Features/Matchmaking.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: d4d01721a4acbec449743b4caa9a8f4a
-folderAsset: yes
-timeCreated: 1517326973
-licenseType: Pro
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
