--- Assets/Scripts/GalaxyManager/Features/Matchmaking/LobbyBrowsing.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: f4a14c3bbd6c2de45b84b8015ce5aac0
-timeCreated: 1517326993
-licenseType: Pro
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
