--- Assets/Scripts/GalaxyManager/Features/Matchmaking/LobbyChat.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: f7dbeda3b2e372349a87cd9fb80db26f
-timeCreated: 1521643189
-licenseType: Pro
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
