--- Assets/Scripts/GalaxyManager/Features/Matchmaking/LobbyCreation.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 7e62c06aadcdc3a469f4f5f49b15b533
-timeCreated: 1517326981
-licenseType: Pro
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
