--- Assets/Scripts/GalaxyManager/Features/Matchmaking/LobbyManagementInGame.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: d36d4f6a42526a3418a5cbe646c94d47
-timeCreated: 1517327023
-licenseType: Pro
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
