--- Assets/Scripts/GalaxyManager/Features/Matchmaking/LobbyManagementMainMenu.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 823ac7783763bb3499ce025e4aec53e3
-timeCreated: 1517327005
-licenseType: Pro
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
