+++ Assets/Scripts/UI/GameMenu/ChatGameController.cs
@@ -20,7 +20,7 @@ public class ChatGameController : MonoBehaviour {
 
     void Update()
     {
-        lobbyMessageHistory = GalaxyManager.Instance.Matchmaking.LobbyChat.lobbyMessageHistory;
+        lobbyMessageHistory = GalaxyManager.Instance.Matchmaking.lobbyMessageHistory;
         if (lobbyMessageHistory.Count > messagesEntriesList.Count)
         {
             DisplayChatMessages();
