+++ Assets/Scripts/UI/MainMenu/OnlineController.cs
@@ -6,7 +6,6 @@ public class OnlineController : MonoBehaviour
     void OnEnable()
     {
         GalaxyManager.Instance.StartMatchmaking();
-        GalaxyManager.Instance.Matchmaking.ShutdownAllMatchmakingClasses();
     }
 
 }
